updated widget
