+++ composeApp/src/androidMain/kotlin/com/jsontextfield/departurescreen/widget/config/WidgetConfigViewModel.kt
@@ -1,14 +1,11 @@
 package com.jsontextfield.departurescreen.widget.config
 
-import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jsontextfield.departurescreen.core.data.IGoTrainDataSource
 import com.jsontextfield.departurescreen.core.data.IPreferencesRepository
 import com.jsontextfield.departurescreen.core.entities.Station
-import com.jsontextfield.departurescreen.core.entities.Trip
 import com.jsontextfield.departurescreen.core.ui.SortMode
-import com.jsontextfield.departurescreen.core.ui.theme.lineColours
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -16,8 +13,6 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.datetime.Clock
-import kotlin.time.Duration.Companion.minutes
 
 class WidgetConfigViewModel(
     private val goTrainDataSource: IGoTrainDataSource,
@@ -85,17 +80,4 @@ data class WidgetConfig(
     val selectedStationCode: String? = null,
     val sortMode: SortMode = SortMode.TIME,
     val opacity: Float = 0.8f,
-) {
-    val trips = List(4) {
-        val line = lineColours.keys.random()
-        Trip(
-            id = "$it",
-            code = line,
-            destination = "Destination",
-            platform = "${2 * it + 1}",
-            color = lineColours[line] ?: Color.Gray,
-            departureTime = Clock.System.now() + it.minutes,
-            lastUpdated = Clock.System.now() - 10.minutes,
-        )
-    }
-}
+)
